+++ nbconvert/__init__.py
@@ -1,5 +1,10 @@
 """Utilities for converting notebooks to and from different formats."""
 
-from . import filters, postprocessors, preprocessors, writers
 from ._version import __version__, version_info  # noqa
-from .exporters import *
+
+try:
+    from . import filters, postprocessors, preprocessors, writers
+    from .exporters import *
+except ModuleNotFoundError:
+    # We hit this condition when the package is not yet fully installed.
+    pass