+++ nbconvert/exporters/script.py
@@ -2,8 +2,12 @@
 
 # Copyright (c) Jupyter Development Team.
 # Distributed under the terms of the Modified BSD License.
+import sys
 
-import entrypoints
+if sys.version_info < (3, 10):
+    from importlib_metadata import entry_points
+else:
+    from importlib.metadata import entry_points
 from traitlets import Dict, default
 
 from .base import get_exporter
@@ -32,8 +36,9 @@ class ScriptExporter(TemplateExporter):
         """
         if lang_name not in self._lang_exporters:
             try:
-                Exporter = entrypoints.get_single("nbconvert.exporters.script", lang_name).load()
-            except entrypoints.NoSuchEntryPoint:
+                exporters = entry_points(group="nbconvert.exporters.script")
+                Exporter = [e for e in exporters if e.name == lang_name][0].load()
+            except (KeyError, IndexError):
                 self._lang_exporters[lang_name] = None
             else:
                 # TODO: passing config is wrong, but changing this revealed more complicated issues